TVGUI-17 Rename requests module
