+++ src/Pages/ShowListPage.tsx
@@ -4,7 +4,7 @@ import { DeleteOutlined } from "@ant-design/icons";
 
 import AddShow from "../components/AddShow";
 import { SearchListContext, UserContext } from "../contexts";
-import { removeShowFromList } from "../requests/requests";
+import { removeShowFromList } from "../requests";
 import './styles/ShowList.css';
 
 const ShowListPage = () => {
